Added register superclass
